+++ math/LinearSolver.h
@@ -26,7 +26,7 @@
 #include <stdexcept>
 #include <string>
 #include <boost/scoped_array.hpp>
-#include <stddef.h>
+#include <cstddef>
 #include <cassert>
 
 /**
@@ -150,7 +150,7 @@ void LinearSolver::solve(T const* A, T* X, T const* B, T* tbuffer, size_t* pbuff
     for (size_t y_col = 0; y_col < m_colsBX; ++y_col) {
         size_t virt_row = 0;
         for (; virt_row < m_colsArowsX; ++virt_row) {
-            int const phys_row = perm[virt_row];
+            int const phys_row = static_cast<const int>(perm[virt_row]);
             T right(p_b_col[phys_row]);
 
             // Move already calculated factors to the right side.
@@ -167,7 +167,7 @@ void LinearSolver::solve(T const* A, T* X, T const* B, T* tbuffer, size_t* pbuff
 
         // Continue below the square part (if any).
         for (; virt_row < m_rowsAB; ++virt_row) {
-            int const phys_row = perm[virt_row];
+            int const phys_row = static_cast<const int>(perm[virt_row]);
             T right(p_b_col[phys_row]);
 
             // Move everything to the right side, then verify it's zero.
@@ -191,13 +191,13 @@ void LinearSolver::solve(T const* A, T* X, T const* B, T* tbuffer, size_t* pbuff
     p_y_col = y_data;
     T const* p_lu_last_col = lu_data + (m_colsArowsX - 1) * m_rowsAB;
     for (size_t x_col = 0; x_col < m_colsBX; ++x_col) {
-        for (int virt_row = m_colsArowsX - 1; virt_row >= 0; --virt_row) {
+        for (int virt_row = static_cast<int>(m_colsArowsX - 1); virt_row >= 0; --virt_row) {
             T right(p_y_col[virt_row]);
 
             // Move already calculated factors to the right side.
             T const* p_lu = p_lu_last_col + perm[virt_row];
             // Go right to left, stop at diagonal.
-            for (int lu_col = m_colsArowsX - 1; lu_col > virt_row; --lu_col) {
+            for (int lu_col = static_cast<int>(m_colsArowsX - 1); lu_col > virt_row; --lu_col) {
                 right -= *p_lu * p_x_col[lu_col];
                 p_lu -= m_rowsAB;
             }
